requirements.txt
fastapi
uvicorn[standard]
gunicorn
pandas
python-dotenv
python-docx
requests
docx2pdf